+++ muacrypt/states.py
@@ -10,6 +10,7 @@ crytographic hashes.
 from __future__ import unicode_literals, print_function
 
 import os
+import logging
 from .chainstore import HeadTracker, BlockService, Chain
 from .myattr import (
     v, attr, attrs, attrib, attrib_text, attrib_bytes,
@@ -167,10 +168,11 @@ class PeerState(object):
         if effective_date < self.autocrypt_timestamp:
             return
         if not keydata:
-            if effective_date > self.last_seen:
+            if effective_date >= self.last_seen:
                 self._append_noac_entry(
                     msg_id=msg_id, msg_date=effective_date,
                 )
+                logging.debug("append noac %s", msg_id)
             return
         self._append_ac_entry(
             msg_id=msg_id, msg_date=effective_date,